Delete metodología.php
There are issues with how git handles non-ascii characters in filenames on different operating systems. OS X normally uses NFD, and Linux/Windows NFC. This file will be deleted and added back on OS X with the config setting `core.precomposeunicode true`. Previously this file shows in the `Untracked files` section of `git status` immediately after checkout on OS X because of the difference in precomposed/decomposed characters in this filename.